fix: improve performances and sentry tracing in POST application
